WIP Try to fix noise
The current implementation (I think) leads to a signed integer overflow, which is UB.
Half of the bytes of the randomness buffer also go unused.
This consistently uses uint{8,16}_t's for the rng calculations, dunno yet if it'll have an effect on the sound.

+++ pxtone/pxtnPulse_NoiseBuilder.h
@@ -13,7 +13,7 @@ private:
 
 	bool    _b_init;
 	short*  _p_tables[ pxWAVETYPE_num ];
-	int32_t _rand_buf [ 2 ];
+	uint16_t _rand_buf [ 2 ];
 
 	void  _random_reset();
 	short _random_get  ();
